mcp/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "mcp-minimal"
version = "1.0.0"
description = "MCP Server to run SQL queries and explore schemas in MS Fabric and PostgreSQL"
license-files = ["LICENSE"]
keywords = ["mcp", "llm", "fabric", "postgres", "warehouse", "lakehouse"]
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
    "annotated-types==0.7.0",
    "anyio==4.10.0",
    "arabic-reshaper==3.0.0",
    "asn1crypto==1.5.1",
    "attrs==25.3.0",
    "authlib==1.6.1",
    "azure-core==1.35.0",
    "azure-identity==1.24.0",
    "psycopg[binary]>=3.2.1",
    "certifi==2025.8.3",
    "cffi==1.17.1",
    "charset-normalizer==3.4.3",
    "click==8.2.1",
    "colorama==0.4.6",
    "contourpy==1.3.3",
    "cryptography==45.0.6",
    "cssselect2==0.8.0",
    "cycler==0.12.1",
    "cyclopts==3.22.5",
    "dnspython==2.7.0",
    "docstring-parser==0.17.0",
    "docutils==0.22",
    "email-validator==2.2.0",
    "exceptiongroup==1.3.0",
    "fastmcp==2.11.3",
    "fonttools==4.59.2",
    "h11==0.16.0",
    "html5lib==1.1",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "httpx-sse==0.4.1",
    "idna==3.10",
    "isodate==0.7.2",
    "jinja2==3.1.6",
    "jsonschema==4.25.0",
    "jsonschema-path==0.3.4",
    "jsonschema-specifications==2025.4.1",
    "kiwisolver==1.4.9",
    "lazy-object-proxy==1.11.0",
    "lxml==6.0.1",
    "markdown-it-py==4.0.0",
    "markupsafe==3.0.2",
    "matplotlib==3.10.6",
    "mdurl==0.1.2",
    "more-itertools==10.7.0",
    "msal==1.33.0",
    "msal-extensions==1.3.1",
    "numpy>=1.23.4",
    "openapi-core==0.19.5",
    "openapi-pydantic==0.5.1",
    "openapi-schema-validator==0.6.3",
    "openapi-spec-validator==0.7.2",
    "oscrypto==1.3.0",
    "packaging==25.0",
    "pandas>=2.2.3",
    "parse==1.20.2",
    "pathable==0.4.4",
    "pillow==11.3.0",
    "pip==24.0",
    "pycparser==2.22",
    "pydantic==2.11.7",
    "pydantic-core==2.33.2",
    "pydantic-settings==2.10.1",
    "pygments==2.19.2",
    "pyhanko==0.30.0",
    "pyhanko-certvalidator==0.28.0",
    "pyjwt==2.10.1",
    "pyodbc==5.2.0",
    "pyparsing==3.2.3",
    "pypdf==6.0.0",
    "pyperclip==1.9.0",
    "pyarrow>=18.0.0",
    "python-bidi==0.6.6",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.1.1",
    "python-multipart==0.0.20",
    "playwright==1.55.0",
    "pytz==2025.2",
    "pywin32==311; sys_platform == 'win32'",
    "pyyaml==6.0.2",
    "referencing==0.36.2",
    "reportlab==4.4.3",
    "requests==2.32.4",
    "rfc3339-validator==0.1.4",
    "rich==14.1.0",
    "rich-rst==1.3.1",
    "rpds-py==0.27.0",
    "seaborn==0.13.2",
    "six==1.17.0",
    "sniffio==1.3.1",
    "sqlparse==0.5.3",
    "sse-starlette==3.0.2",
    "starlette==0.47.2",
    "svglib==1.5.1",
    "tinycss2==1.4.0",
    "typing-extensions==4.14.1",
    "typing-inspection==0.4.1",
    "tzdata==2025.2",
    "tzlocal==5.3.1",
    "uritools==5.0.0",
    "urllib3==2.5.0",
    "uvicorn==0.35.0",
    "webencodings==0.5.1",
    "werkzeug==3.1.1",
    "xhtml2pdf==0.2.17",
]

[project.urls]

[project.scripts]

[tool.setuptools]
packages = ["src", "src.services", "src.tools"]
package-dir = {"" = ".", "services" = "src/services", "tools" = "src/tools"}
